refactor: renomeia modulo para RenderStatusByQuarterTag e organiza arquivos
